navigaation kanssa säätämistä
